+++ src/init.c
@@ -29,6 +29,7 @@ extern SEXP _IncDTW_cpp_dtw2vec_v32(SEXP, SEXP);
 extern SEXP _IncDTW_cpp_dtw2vec_ws(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_dtw2vec_ws_ea(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_get_tube(SEXP, SEXP);
+extern SEXP _IncDTW_cpp_get_tube_mv(SEXP, SEXP);
 extern SEXP _IncDTW_cpp_kNN_rev(SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_local_min(SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_norm01(SEXP, SEXP, SEXP, SEXP);
@@ -41,11 +42,13 @@ extern SEXP _IncDTW_cpp_rundtw_znorm_lot(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEX
 extern SEXP _IncDTW_cpp_rundtw_znorm_mv(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_rundtw_znorm_mv_lot(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_set_tube(SEXP, SEXP, SEXP);
-extern SEXP _IncDTW_cpp_set_tube_mv(SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_cpp_znorm(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_GCM_cpp(SEXP, SEXP);
 extern SEXP _IncDTW_GCM_Sakoe_cpp(SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_get_lb(SEXP, SEXP, SEXP, SEXP);
+extern SEXP _IncDTW_get_lb_mv1(SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP _IncDTW_get_lb_mv2(SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP _IncDTW_get_lb_mv22(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_get_lb_znorm(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_IGCM_cpp(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _IncDTW_IGCM_Sakoe_cpp(SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -77,6 +80,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_IncDTW_cpp_dtw2vec_ws",          (DL_FUNC) &_IncDTW_cpp_dtw2vec_ws,           4},
     {"_IncDTW_cpp_dtw2vec_ws_ea",       (DL_FUNC) &_IncDTW_cpp_dtw2vec_ws_ea,        5},
     {"_IncDTW_cpp_get_tube",            (DL_FUNC) &_IncDTW_cpp_get_tube,             2},
+    {"_IncDTW_cpp_get_tube_mv",         (DL_FUNC) &_IncDTW_cpp_get_tube_mv,          2},
     {"_IncDTW_cpp_kNN_rev",             (DL_FUNC) &_IncDTW_cpp_kNN_rev,              3},
     {"_IncDTW_cpp_local_min",           (DL_FUNC) &_IncDTW_cpp_local_min,            3},
     {"_IncDTW_cpp_norm01",              (DL_FUNC) &_IncDTW_cpp_norm01,               4},
@@ -89,11 +93,13 @@ static const R_CallMethodDef CallEntries[] = {
     {"_IncDTW_cpp_rundtw_znorm_mv",     (DL_FUNC) &_IncDTW_cpp_rundtw_znorm_mv,     11},
     {"_IncDTW_cpp_rundtw_znorm_mv_lot", (DL_FUNC) &_IncDTW_cpp_rundtw_znorm_mv_lot, 15},
     {"_IncDTW_cpp_set_tube",            (DL_FUNC) &_IncDTW_cpp_set_tube,             3},
-    {"_IncDTW_cpp_set_tube_mv",         (DL_FUNC) &_IncDTW_cpp_set_tube_mv,          3},
     {"_IncDTW_cpp_znorm",               (DL_FUNC) &_IncDTW_cpp_znorm,                4},
     {"_IncDTW_GCM_cpp",                 (DL_FUNC) &_IncDTW_GCM_cpp,                  2},
     {"_IncDTW_GCM_Sakoe_cpp",           (DL_FUNC) &_IncDTW_GCM_Sakoe_cpp,            3},
     {"_IncDTW_get_lb",                  (DL_FUNC) &_IncDTW_get_lb,                   4},
+    {"_IncDTW_get_lb_mv1",              (DL_FUNC) &_IncDTW_get_lb_mv1,               5},
+    {"_IncDTW_get_lb_mv2",              (DL_FUNC) &_IncDTW_get_lb_mv2,               5},
+    {"_IncDTW_get_lb_mv22",             (DL_FUNC) &_IncDTW_get_lb_mv22,              5},
     {"_IncDTW_get_lb_znorm",            (DL_FUNC) &_IncDTW_get_lb_znorm,             7},
     {"_IncDTW_IGCM_cpp",                (DL_FUNC) &_IncDTW_IGCM_cpp,                 4},
     {"_IncDTW_IGCM_Sakoe_cpp",          (DL_FUNC) &_IncDTW_IGCM_Sakoe_cpp,           5},