+++ NEWS.md
@@ -1,4 +1,23 @@
-### Version 1.1.2.03 (in dev)
+### Version 1.1.4.01
+
+##### Bug Fixes
+
+- there was a buggy test "Double Incremental Matrix EQUAL Scratch" in the test file "test_dtw.R". The error ocurred for some random seeds when the optimal warping path is not unique. The DTW distance measure and the GCM is always unique and correct, but the direction matrix and warping path are not. I set a seed, so that the error should not happen again in this test file. 
+-...
+
+##### New Features
+
+- new function: lb_keogh() and lb_keogh_tube()
+- ...
+
+##### other changes
+
+- I reduced the file size of the vignette "Theory and Applications for the RPackage IncDTW.pdf" by about 800Kb.
+- ...
+
+
+
+### Version 1.1.3
 
 ##### Bug Fixes
 
@@ -7,7 +26,7 @@
 
 ##### New Features
 
-- New Vignettes
+- New vignette that replaces the old vignettes
 - replaced the function norm() by scale(), same functionality. norm() still works, but prints a warning to be deprecated.
 - replaced the arguments 'normalize' by 'scale' for the function rundtw(). See details of the function documentation.